airflow-basics/requirements.txt
apache-airflow==2.9.1
